requirements.txt
conllu
numpy
python-dotenv
